requirements.txt
altgraph==0.16.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
configparser==4.0.2
cryptography==2.8
decorator==4.4.1
django-frontend==1.8.0
etelemetry==0.1.2
filelock==3.0.12
fitz==0.0.1.dev2
flasgger==0.9.3
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
httplib2==0.15.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
lxml==4.4.2
MarkupSafe==1.1.1
mistune==0.8.4
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.0
nipype==1.4.0
numpy==1.18.1
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
packaging==20.0
pandas==0.25.3
pdfminer==20191125
pefile==2019.4.18
Pillow==6.2.1
prov==1.5.3
pycparser==2.19
pycryptodome==3.9.4
pydot==1.4.1
pydotplus==2.0.2
PyInstaller==3.5
PyMuPDF==1.16.10
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
pytesseract==0.3.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pyxnat==1.2.1.0.post3
PyYAML==5.3
qrcode==6.1
rdflib==4.2.2
reportlab==3.5.32
requests==2.22.0
scipy==1.4.1
simplejson==3.17.0
six==1.13.0
traits==5.2.0
urllib3==1.25.7
Werkzeug==0.16.0
